initial implementation of web-server
