+++ packages/next/server/base-server.ts
@@ -1,5 +1,5 @@
 import type { __ApiPreviewProps } from './api-utils'
-import type { CustomRoutes, Header } from '../lib/load-custom-routes'
+import type { CustomRoutes } from '../lib/load-custom-routes'
 import type { DomainLocale } from './config'
 import type { DynamicRoutes, PageChecker, Params, Route } from './router'
 import type { FontManifest } from './font-utils'
@@ -7,10 +7,8 @@ import type { LoadComponentsReturnType } from './load-components'
 import type { MiddlewareManifest } from '../build/webpack/plugins/middleware-plugin'
 import type { NextConfig, NextConfigComplete } from './config-shared'
 import type { NextParsedUrlQuery, NextUrlWithParsedQuery } from './request-meta'
-import type { ParsedUrl } from '../shared/lib/router/utils/parse-url'
 import type { ParsedUrlQuery } from 'querystring'
-import type { PrerenderManifest } from '../build'
-import type { Redirect, Rewrite, RouteType } from '../lib/load-custom-routes'
+import type { Rewrite } from '../lib/load-custom-routes'
 import type { RenderOpts, RenderOptsPartial } from './render'
 import type { ResponseCacheEntry, ResponseCacheValue } from './response-cache'
 import type { UrlWithParsedQuery } from 'url'
@@ -20,9 +18,9 @@ import type { PagesManifest } from '../build/webpack/plugins/pages-manifest-plug
 import type { BaseNextRequest, BaseNextResponse } from './base-http'
 
 import { join, resolve } from 'path'
-import { parse as parseQs, stringify as stringifyQs } from 'querystring'
+import { parse as parseQs } from 'querystring'
 import { format as formatUrl, parse as parseUrl } from 'url'
-import { getRedirectStatus, modifyRouteRegex } from '../lib/load-custom-routes'
+import { getRedirectStatus } from '../lib/load-custom-routes'
 import {
   SERVERLESS_DIRECTORY,
   SERVER_DIRECTORY,
@@ -39,12 +37,7 @@ import * as envConfig from '../shared/lib/runtime-config'
 import { DecodeError, normalizeRepeatedSlashes } from '../shared/lib/utils'
 import { setLazyProp, getCookieParser, tryGetPreviewData } from './api-utils'
 import { isTargetLikeServerless } from './utils'
-import pathMatch from '../shared/lib/router/utils/path-match'
 import Router, { replaceBasePath, route } from './router'
-import {
-  compileNonPath,
-  prepareDestination,
-} from '../shared/lib/router/utils/prepare-destination'
 import { PayloadOptions, setRevalidateHeaders } from './send-payload'
 import { IncrementalCache } from './incremental-cache'
 import { execOnce } from '../shared/lib/utils'
@@ -63,8 +56,8 @@ import { parseNextUrl } from '../shared/lib/router/utils/parse-next-url'
 import isError, { getProperError } from '../lib/is-error'
 import { MIDDLEWARE_ROUTE } from '../lib/constants'
 import { addRequestMeta, getRequestMeta } from './request-meta'
-
-const getCustomRouteMatcher = pathMatch(true)
+import { createHeaderRoute, createRedirectRoute } from './server-route-utils'
+import { PrerenderManifest } from '../build'
 
 export type FindComponentsResult = {
   components: LoadComponentsReturnType
@@ -185,6 +178,15 @@ export default abstract class Server {
   protected abstract generateStaticRotes(): Route[]
   protected abstract generateFsStaticRoutes(): Route[]
   protected abstract generateCatchAllMiddlewareRoute(): Route | undefined
+  protected abstract generateRewrites({
+    restrictedRedirectPaths,
+  }: {
+    restrictedRedirectPaths: string[]
+  }): {
+    beforeFiles: Route[]
+    afterFiles: Route[]
+    fallback: Route[]
+  }
   protected abstract getFilesystemPaths(): Set<string>
   protected abstract getMiddleware(): {
     match: (pathname: string | null | undefined) =>
@@ -238,27 +240,10 @@ export default abstract class Server {
     renderOpts: RenderOpts
   ): Promise<RenderResult | null>
 
-  protected abstract streamResponseChunk(
-    res: BaseNextResponse,
-    chunk: any
-  ): void
-
   protected abstract handleCompression(
     req: BaseNextRequest,
     res: BaseNextResponse
   ): void
-
-  protected abstract proxyRequest(
-    req: BaseNextRequest,
-    res: BaseNextResponse,
-    parsedUrl: ParsedUrl
-  ): Promise<{ finished: boolean }>
-
-  protected abstract imageOptimizer(
-    req: BaseNextRequest,
-    res: BaseNextResponse,
-    parsedUrl: UrlWithParsedQuery
-  ): Promise<{ finished: boolean }>
 
   protected abstract loadEnvConfig(params: { dev: boolean }): void
 
@@ -742,159 +727,24 @@ export default abstract class Server {
       ...staticFilesRoutes,
     ]
 
-    const restrictedRedirectPaths = ['/_next'].map((p) =>
-      this.nextConfig.basePath ? `${this.nextConfig.basePath}${p}` : p
-    )
-
-    const getCustomRoute = (
-      r: Rewrite | Redirect | Header,
-      type: RouteType
-    ) => {
-      const match = getCustomRouteMatcher(
-        r.source,
-        !(r as any).internal
-          ? (regex: string) =>
-              modifyRouteRegex(
-                regex,
-                type === 'redirect' ? restrictedRedirectPaths : undefined
-              )
-          : undefined
-      )
-
-      return {
-        ...r,
-        type,
-        match,
-        name: type,
-        fn: async (_req, _res, _params, _parsedUrl) => ({ finished: false }),
-      } as Route & Rewrite & Header
-    }
+    const restrictedRedirectPaths = this.nextConfig.basePath
+      ? [`${this.nextConfig.basePath}/_next`]
+      : ['/_next']
 
     // Headers come very first
     const headers = this.minimalMode
       ? []
-      : this.customRoutes.headers.map((r) => {
-          const headerRoute = getCustomRoute(r, 'header')
-          return {
-            match: headerRoute.match,
-            has: headerRoute.has,
-            type: headerRoute.type,
-            name: `${headerRoute.type} ${headerRoute.source} header route`,
-            fn: async (_req, res, params, _parsedUrl) => {
-              const hasParams = Object.keys(params).length > 0
-
-              for (const header of (headerRoute as Header).headers) {
-                let { key, value } = header
-                if (hasParams) {
-                  key = compileNonPath(key, params)
-                  value = compileNonPath(value, params)
-                }
-                res.setHeader(key, value)
-              }
-              return { finished: false }
-            },
-          } as Route
-        })
+      : this.customRoutes.headers.map((rule) =>
+          createHeaderRoute({ rule, restrictedRedirectPaths })
+        )
 
     const redirects = this.minimalMode
       ? []
-      : this.customRoutes.redirects.map((redirect) => {
-          const redirectRoute = getCustomRoute(redirect, 'redirect')
-          return {
-            internal: redirectRoute.internal,
-            type: redirectRoute.type,
-            match: redirectRoute.match,
-            has: redirectRoute.has,
-            statusCode: redirectRoute.statusCode,
-            name: `Redirect route ${redirectRoute.source}`,
-            fn: async (req, res, params, parsedUrl) => {
-              const { parsedDestination } = prepareDestination({
-                appendParamsToQuery: false,
-                destination: redirectRoute.destination,
-                params: params,
-                query: parsedUrl.query,
-              })
-
-              const { query } = parsedDestination
-              delete (parsedDestination as any).query
-
-              parsedDestination.search = stringifyQuery(req, query)
-
-              let updatedDestination = formatUrl(parsedDestination)
-
-              if (updatedDestination.startsWith('/')) {
-                updatedDestination =
-                  normalizeRepeatedSlashes(updatedDestination)
-              }
-
-              res
-                .redirect(
-                  updatedDestination,
-                  getRedirectStatus(redirectRoute as Redirect)
-                )
-                .body(updatedDestination)
-                .send()
-
-              return {
-                finished: true,
-              }
-            },
-          } as Route
-        })
-
-    const buildRewrite = (rewrite: Rewrite, check = true) => {
-      const rewriteRoute = getCustomRoute(rewrite, 'rewrite')
-      return {
-        ...rewriteRoute,
-        check,
-        type: rewriteRoute.type,
-        name: `Rewrite route ${rewriteRoute.source}`,
-        match: rewriteRoute.match,
-        fn: async (req, res, params, parsedUrl) => {
-          const { newUrl, parsedDestination } = prepareDestination({
-            appendParamsToQuery: true,
-            destination: rewriteRoute.destination,
-            params: params,
-            query: parsedUrl.query,
-          })
-
-          // external rewrite, proxy it
-          if (parsedDestination.protocol) {
-            return this.proxyRequest(req, res, parsedDestination)
-          }
-
-          addRequestMeta(req, '_nextRewroteUrl', newUrl)
-          addRequestMeta(req, '_nextDidRewrite', newUrl !== req.url)
-
-          return {
-            finished: false,
-            pathname: newUrl,
-            query: parsedDestination.query,
-          }
-        },
-      } as Route
-    }
-
-    let beforeFiles: Route[] = []
-    let afterFiles: Route[] = []
-    let fallback: Route[] = []
-
-    if (!this.minimalMode) {
-      if (Array.isArray(this.customRoutes.rewrites)) {
-        afterFiles = this.customRoutes.rewrites.map((r) => buildRewrite(r))
-      } else {
-        beforeFiles = this.customRoutes.rewrites.beforeFiles.map((r) =>
-          buildRewrite(r, false)
+      : this.customRoutes.redirects.map((rule) =>
+          createRedirectRoute({ rule, restrictedRedirectPaths })
         )
-        afterFiles = this.customRoutes.rewrites.afterFiles.map((r) =>
-          buildRewrite(r)
-        )
-        fallback = this.customRoutes.rewrites.fallback.map((r) =>
-          buildRewrite(r)
-        )
-      }
-    }
 
+    const rewrites = this.generateRewrites({ restrictedRedirectPaths })
     const catchAllMiddleware = this.generateCatchAllMiddlewareRoute()
 
     const catchAllRoute: Route = {
@@ -968,11 +818,7 @@ export default abstract class Server {
     return {
       headers,
       fsRoutes,
-      rewrites: {
-        beforeFiles,
-        afterFiles,
-        fallback,
-      },
+      rewrites,
       redirects,
       catchAllRoute,
       catchAllMiddleware,
@@ -1962,23 +1808,7 @@ export function prepareServerlessUrl(
   })
 }
 
-// since initial query values are decoded by querystring.parse
-// we need to re-encode them here but still allow passing through
-// values from rewrites/redirects
-export const stringifyQuery = (req: BaseNextRequest, query: ParsedUrlQuery) => {
-  const initialQueryValues = Object.values(
-    getRequestMeta(req, '__NEXT_INIT_QUERY') || {}
-  )
-
-  return stringifyQs(query, undefined, undefined, {
-    encodeURIComponent(value) {
-      if (initialQueryValues.some((val) => val === value)) {
-        return encodeURIComponent(value)
-      }
-      return value
-    },
-  })
-}
+export { stringifyQuery } from './server-route-utils'
 
 class NoFallbackError extends Error {}
