+++ packages/next/server/web-server.ts
@@ -1,11 +1,55 @@
-import BaseServer, {
-  Options,
-  FindComponentsResult,
-  prepareServerlessUrl,
-  stringifyQuery,
-} from './base-server'
+import type { WebNextRequest, WebNextResponse } from './base-http'
+import type { RenderOpts } from './render'
+import type RenderResult from './render-result'
+import type { NextParsedUrlQuery } from './request-meta'
+import type { Params } from './router'
+import type { PayloadOptions } from './send-payload'
+import type { PrerenderManifest } from '../build'
+
+import BaseServer, { Options } from './base-server'
 
 export default class NextWebServer extends BaseServer {
+  constructor(
+    options: Options & {
+      prerenderManifest: PrerenderManifest
+    }
+  ) {
+    super(options)
+  }
+  protected generateRewrites() {
+    // @TODO: assuming minimal mode right now
+    return {
+      beforeFiles: [],
+      afterFiles: [],
+      fallback: [],
+    }
+  }
+  protected handleCompression() {
+    // @TODO
+  }
+  protected getRoutesManifest() {
+    return {
+      headers: [],
+      rewrites: {
+        fallback: [],
+        afterFiles: [],
+        beforeFiles: [],
+      },
+      redirects: [],
+    }
+  }
+  protected getPagePath() {
+    // @TODO
+    return ''
+  }
+  protected getPublicDir() {
+    // @TODO
+    return ''
+  }
+  protected getBuildId() {
+    // @TODO
+    return ''
+  }
   protected loadEnvConfig() {
     // @TODO
   }
@@ -45,5 +89,56 @@ export default class NextWebServer extends BaseServer {
   }
   protected getFilesystemPaths() {
     return new Set<string>()
+  }
+  protected async renderHTML(
+    req: WebNextRequest,
+    res: WebNextResponse,
+    pathname: string,
+    query: NextParsedUrlQuery,
+    renderOpts: RenderOpts
+  ): Promise<RenderResult | null> {
+    // @TODO
+    console.log(pathname)
+    return null
+  }
+  protected async sendRenderResult(
+    req: WebNextRequest,
+    res: WebNextResponse,
+    options: {
+      result: RenderResult
+      type: 'html' | 'json'
+      generateEtags: boolean
+      poweredByHeader: boolean
+      options?: PayloadOptions | undefined
+    }
+  ): Promise<void> {
+    // @TODO
+    console.log(options)
+  }
+  protected async runApi() {
+    // @TODO
+    return true
+  }
+  protected async findPageComponents(
+    pathname: string,
+    query?: NextParsedUrlQuery,
+    params?: Params | null
+  ) {
+    // @TODO
+    return null
+  }
+  protected getPrerenderManifest() {
+    // @TODO
+    return {
+      version: 3 as const,
+      routes: {},
+      dynamicRoutes: {},
+      notFoundRoutes: [],
+      preview: {
+        previewModeId: '',
+        previewModeEncryptionKey: '',
+        previewModeSigningKey: '',
+      },
+    }
   }
 }
